Remove all references to Autofac from the solution in favor of ReactiveUI Splat for DI.
Original plan was to use Autofac for IoC, but due to unexpected surprises while working with ReactiveUI for the first time, I’ve abandoned the original plan and instead will be using ReactiveUI’s native Splat instead.

ReactiveUI provides documentation on how to adapt its framework to be usable with Autofac, but the documentation is at least four years out of date, as it uses parts of Autofac that have been marked as obsolete over four years ago.

Due to the scope of this application, manually replacing Splat for Autofac without any proper guidance is probably not a very good time investment as it wouldn’t really do anything for the application that Splat isn’t capable of, and there’s probably something to gain by using ReactiveUI’s favored IoC implementation. :)