Add a small margin off-set to the close window button in the titlebar to allign it with minimize button and center them slightly more
The minimze and close window buttons were slightly un-even relative to eachother, and this change applies a negative margin to make the elements allign to eachother.
This also applies equal negative margin for both of them to make it more centered in the title bar, as it was slightly off-center on the verticle axis.