Replace the inbuilt binding list type over from List<T> to ObservableCollection<T> in order to update the view when items are removed or added to the Binding Cache loaded in memory.
Because of weirdness with binding List<T> to view properties, any attempt to add or remove items doesn't update the view because the event is being raised on the wrong thread.

IE: Removing Checklist 1, and it still remains in the list.  Replacing List<T> with ObservableCollection<T> is the easiest way to fix this.  Alternative fix to this issue would be to use Dispatcher.Invoke to raise the events on the main thread, but it's far cleaner to use the more adaptive collection type.