Update the NewDefaultChecklistAsync and DeleteListAsync methods to make the experience of using it with a slow DB connection a lot better.
Instead of waiting for the database to refresh after modifying anything in it, the methods instead modify local collection and send orders to service class to update the DB behind the scenes.
This makes the application a lot more usable if the DB connection is being slow or if the database is far away.