Replace usage of System.Windows.Media library in favor of hex strings and improve documentation in SettingsViewModel
Previously SettingsViewModel used System.Windows.Media library to deal with colors in WPF view, but this library is WPF-only.

In order to not rely on WPF libraries in the back-end, I've opted for using a string with hex value instead of Brush.Color, and updated the references to it.
WPF has inbuilt string to hex value converter, so explicitly converting is not necessary.